Marking Events. Social Page Design Rework Began. More.
Marking of events with the star button has been added. The algorithm for determining the active class has been altered. The design of the socials page is almost done being modernized. The CCHS website has been added to the socials page. The method of adding classes has been made to match that of events.